update: use combineReducers
