+++ src/index.tsx
@@ -4,18 +4,18 @@ import { Provider } from "react-redux";
 import { createStore } from "redux";
 
 import App from "./App";
-import postReducer from "./reducer";
+import { rootReducer } from "./reducer";
 import * as serviceWorker from "./serviceWorker";
 
 import "./index.css";
 
-const store = createStore(postReducer);
+const store = createStore(rootReducer);
 
 ReactDOM.render(
   <Provider store={store}>
     <App />
   </Provider>,
-  document.getElementById("root"),
+  document.getElementById("root")
 );
 
 // If you want your app to work offline and load faster, you can change
